image-caption-generator-main/backend/requirements.txt
Flask
Flask-PyMongo
Flask-CORS
bcrypt
python-dotenv
transformers
torch
Pillow
pymongo
accelerate
numpy
requests 
google-generativeai